fix(term): ensure terminal attr index doesn't exceed `TERM_ATTRS_MAX`
Problem: Currently terminal highlight attribute buffers are statically allocated
be the size of `TERM_ATTRS_MAX`. This unique case isn't respected in
some places in the ui_compositor. Due to this, when a terminal window
has lines longer them `TERM_ATTRS_MAX`, the compositor will go past the
end of the buffer causing a crash due to out of bounds access.

Solution: Pass this `max_attrs` value down to the compositor's line drawing code
so that we can prevent going out of bounds. This will be set to
`TERM_ATTRS_MAX` for terminals, and 0 if there is no limit.

Fixes #30374